+++ Utils/src/NeuralNet.cc
@@ -9,9 +9,9 @@ NeuralNet::NeuralNet(unsigned int in, unsigned int out):
 }
 
 NeuralNet::~NeuralNet() { 
-  for (float *layer : layers)
+  for (double *layer : layers)
     delete[] layer;
-  for (float *b : bs) {
+  for (double *b : bs) {
     delete[] b;
   }
   for (unsigned int j=0; j!=nIn; ++j) {
@@ -22,12 +22,12 @@ NeuralNet::~NeuralNet() {
       delete[]Ws[i][j];
     }
   }
-  for (float **W : Ws) {
+  for (double **W : Ws) {
     delete[] W;
   }
 }
 
-void NeuralNet::AddLayer(unsigned int in, unsigned int out, float **W, float *b, bool isFinal) {
+void NeuralNet::AddLayer(unsigned int in, unsigned int out, double **W, double *b, bool isFinal) {
   unsigned int previousLayerSize = (hiddenLayerSizes.size()==0) ? nIn : hiddenLayerSizes.back();
   assert(previousLayerSize==in);
   if (isFinal)
@@ -40,7 +40,7 @@ void NeuralNet::AddLayer(unsigned int in, unsigned int out, float **W, float *b,
   integrityChecked = false;
 }
 
-void NeuralNet::AddBranchAddress(float *input, float mean, float stdev, const char *name/*=""*/) {
+void NeuralNet::AddBranchAddress(float *input, double mean, double stdev, const char *name/*=""*/) {
   inputs.push_back(input);
   mus.push_back(mean);
   sigmas.push_back(stdev);
@@ -51,14 +51,14 @@ void NeuralNet::AddBranchAddress(float *input, float mean, float stdev, const ch
 void NeuralNet::AllocateMemory() {
   // pre-allocate memory for linear operations
   // faster than using TMatrix, which allocates memory for each operation
-  float *tmpLayer;
+  double *tmpLayer;
   for (unsigned int iL=0; iL!=hiddenLayerSizes.size(); ++iL) {
     fprintf(stderr,"allocating layer of size %i\n",(int)hiddenLayerSizes[iL]);
-    tmpLayer = new float[hiddenLayerSizes[iL]];
+    tmpLayer = new double[hiddenLayerSizes[iL]];
     layers.push_back(tmpLayer);
   }
   fprintf(stderr,"allocating layer of size %i\n",(int)nOut);
-  tmpLayer = new float[nOut];
+  tmpLayer = new double[nOut];
   layers.push_back(tmpLayer);
   integrityChecked = false;
 }
@@ -84,20 +84,20 @@ bool NeuralNet::CheckIntegrity() const {
   return true;
 }
 
-float *NeuralNet::Evaluate() const {
+double *NeuralNet::Evaluate() const {
 	if (!integrityChecked)
   	assert(CheckIntegrity());
   for (unsigned int iOut=0; iOut!=hiddenLayerSizes[0]; ++iOut) {
-    float outVal = bs[0][iOut];
+    double outVal = bs[0][iOut];
     for (unsigned int iIn=0; iIn!=nIn; ++iIn) {
-      float tmpIn = ( *(inputs[iIn]) - mus[iIn] ) / sigmas[iIn];
+      double tmpIn = ( *(inputs[iIn]) - mus[iIn] ) / sigmas[iIn];
       outVal += tmpIn * Ws[0][iIn][iOut];
     }
     layers[0][iOut] = TMath::TanH(outVal);
   }
   for (unsigned int iLayer=1; iLayer!=hiddenLayerSizes.size(); ++iLayer) {
    for (unsigned int iOut=0; iOut!=hiddenLayerSizes[iLayer]; ++iOut) {
-      float outVal = bs[iLayer][iOut];
+      double outVal = bs[iLayer][iOut];
       for (unsigned int iIn=0; iIn!=hiddenLayerSizes[iLayer-1]; ++iIn) {
         outVal += layers[iLayer-1][iIn] * Ws[iLayer][iIn][iOut];
         
@@ -105,12 +105,12 @@ float *NeuralNet::Evaluate() const {
       layers[iLayer][iOut] = TMath::TanH(outVal);
     } 
   }
-  float Z = 0;
-  float *outLayer = layers.back();
+  double Z = 0;
+  double *outLayer = layers.back();
   unsigned int nextToLast = hiddenLayerSizes.size();
-  float *prevLayer = layers[nextToLast-1];
+  double *prevLayer = layers[nextToLast-1];
   for (unsigned int iOut=0; iOut!=nOut; ++iOut) {
-    float outVal = bs[nextToLast][iOut];
+    double outVal = bs[nextToLast][iOut];
     for (unsigned int iIn=0; iIn!=hiddenLayerSizes.back(); ++iIn) {
       outVal += prevLayer[iIn] * Ws[nextToLast][iIn][iOut];
     }
